requirements.txt
pygame==1.9.4
pyserial==3.4